ushort To uint.

+++ Reluca.Tests/Evaluates/FeaturePatternExtractorTest.cs
@@ -21,9 +21,9 @@ namespace Reluca.Tests.Evaluates
             var context = CreateBoardContext(1, 2, ResourceType.In);
             Target.Initialize(resource, new NoneNormalizer());
             var actual = Target.Extract(context);
-            Assert.AreEqual(RadixHelper.ToUInt16("2011111021", 3), actual[FeaturePattern.Type.Edge2X][0]);
-            Assert.AreEqual(RadixHelper.ToUInt16("2110111202", 3), actual[FeaturePattern.Type.Edge2X][1]);
-            Assert.AreEqual(RadixHelper.ToUInt16("2110110112", 3), actual[FeaturePattern.Type.Corner2X5][0]);
+            Assert.AreEqual(RadixHelper.ToUInt32("2011111021", 3), actual[FeaturePattern.Type.Edge2X][0]);
+            Assert.AreEqual(RadixHelper.ToUInt32("2110111202", 3), actual[FeaturePattern.Type.Edge2X][1]);
+            Assert.AreEqual(RadixHelper.ToUInt32("2110110112", 3), actual[FeaturePattern.Type.Corner2X5][0]);
         }
 
         [TestMethod]
@@ -33,12 +33,12 @@ namespace Reluca.Tests.Evaluates
             var context = CreateBoardContext(1, 2, ResourceType.In);
             Target.Initialize(resource, new FeaturePatternNormalizer());
             var actual = Target.Extract(context);
-            // Assert.AreEqual(RadixHelper.ToUInt16("2011111021", 3), actual[FeaturePattern.Type.Edge2X][0]);
-            Assert.AreEqual(RadixHelper.ToUInt16("0211111201", 3), actual[FeaturePattern.Type.Edge2X][0]);
-            // Assert.AreEqual(RadixHelper.ToUInt16("2110111202", 3), actual[FeaturePattern.Type.Edge2X][1]);
-            Assert.AreEqual(RadixHelper.ToUInt16("0112111020", 3), actual[FeaturePattern.Type.Edge2X][1]);
-            // Assert.AreEqual(RadixHelper.ToUInt16("2110110112", 3), actual[FeaturePattern.Type.Corner2X5][0]);
-            Assert.AreEqual(RadixHelper.ToUInt16("0112112110", 3), actual[FeaturePattern.Type.Corner2X5][0]);
+            // Assert.AreEqual(RadixHelper.ToUInt32("2011111021", 3), actual[FeaturePattern.Type.Edge2X][0]);
+            Assert.AreEqual(RadixHelper.ToUInt32("0211111201", 3), actual[FeaturePattern.Type.Edge2X][0]);
+            // Assert.AreEqual(RadixHelper.ToUInt32("2110111202", 3), actual[FeaturePattern.Type.Edge2X][1]);
+            Assert.AreEqual(RadixHelper.ToUInt32("0112111020", 3), actual[FeaturePattern.Type.Edge2X][1]);
+            // Assert.AreEqual(RadixHelper.ToUInt32("2110110112", 3), actual[FeaturePattern.Type.Corner2X5][0]);
+            Assert.AreEqual(RadixHelper.ToUInt32("0112112110", 3), actual[FeaturePattern.Type.Corner2X5][0]);
         }
     }
 }
